[STYLE] CSS 전역에 지정된 내용들 수정
