+++ src/App.js
@@ -35,7 +35,6 @@ import SelectCategory from './pages/main/SelectCategory';
 import ReviewList from './pages/admin/member/ReviewList';
 import SelectCity from './pages/main/SelectCity';
 import ApplyDetail from './pages/admin/business/ApplyDetail';
-import MySavedPlace from './pages/mypage/MySavedPlace';
 
 
 function App() {
@@ -73,8 +72,8 @@ function App() {
         <Route path='mypage/myreviewlist' element={<Layout_MyPage/>}>
           <Route index element={<MyReviewList/>}/>
         </Route>
-        <Route path='mypage/mysavedplace' element={<Layout_MyPage/>}>
-          <Route index element={<MySavedPlace/>}/>
+        <Route path='mypage/bookmark' element={<Layout_MyPage/>}>
+          <Route index element={<Bookmark/>}/>
         </Route>
