+++ src/layouts/Layout_Mypage.css
@@ -1,12 +1,9 @@
-html, body {
-    height: 100%;
-    margin: 0;
-}
-
 .layout-mypage {
     display: flex;
     flex-direction: column;
     min-height: 100vh;
+    height: 100%;
+    margin: 0;
     font-family: "Noto Serif KR", Times New Roman;
 }
