+++ src/pages/mypage/Bookmark.css
@@ -1,4 +1,4 @@
-.savedplacelist-container {
+.bookmark-container {
     background-color: #FFFDD0;
     padding: 16px;
 }
@@ -17,7 +17,7 @@
     border-radius: 20px;
 }
 
-.savedplacelist {
+.bookmark {
     display: flex;
     flex-direction: column;
     gap: 20px;
