+++ src/pages/mypage/ChangePassword.css
@@ -30,7 +30,7 @@
     text-align: left;
 }
 
-form {
+.changepassword-form {
     display: flex;
     flex-direction: column;
     flex: 1;
