+++ src/pages/mypage/MyBusinessDetail.css
@@ -8,26 +8,26 @@
     margin-bottom: 20px;
 }
 
-label {
+.profile-pic-label, .businessinfo-label {
     display: block;
     margin-bottom: 5px;
 }
 
-input[type="text"],
-input[type="tel"],
-input[type="url"],
-textarea {
+.mybusinessdetail-body .form-group input[type="text"],
+.mybusinessdetail-body .form-group input[type="tel"],
+.mybusinessdetail-body .form-group input[type="url"],
+.mybusinessdetail-body .form-group textarea {
     width: 100%;
     padding: 10px;
     border: 1px solid #ddd;
     border-radius: 4px;
 }
 
-textarea {
+.mybusinessdetail-body .form-group textarea {
     height: 100px;
 }
 
-button[type="submit"] {
+.mybusinessdetail-body .form-actions button[type="submit"] {
     background-color: #E3F988;
     color: #6a8346;
     padding: 10px 20px;
@@ -36,6 +36,7 @@ button[type="submit"] {
     cursor: pointer;
     border: 1px solid #bbb;
 }
+
 
 .add-image-button {
     border: 1px solid #bbb;
@@ -54,7 +55,7 @@ button[type="submit"] {
     object-fit: cover;  /* 이미지 비율 유지 */
 }
 
-button[type="submit"]:hover {
+.mybusinessdetail-body .form-actions button[type="submit"]:hover {
     background-color: #d2e877;
 }
